requirements.txt
pandas==1.5.3
numpy==1.23.5
prophet==1.1.4
kafka-python==2.0.2
psycopg2-binary==2.9.6
scikit-learn==1.2.0
cmdstanpy==1.1.0
boto3==1.26.0
plotly==5.13.0
kaleido==0.1.0
matplotlib==3.6.0